Move sesame2spiner files into a library and executable src/ directories. Modify cmake so that we build a library.
