+++ sesame2spiner/sesame2spiner/parser.hpp
@@ -21,6 +21,48 @@
 #include <unordered_map>
 #include <vector>
 
+const std::string EXAMPLESTRING = R"(
+# air.dat
+# These are comments. 
+# The "#" character must be at the beginning of a line.
+# only matid is required. All others override defaults.
+matid = 5030
+name = air
+# rho is in g/cm^3
+rhomin = 1e-2
+rhomax = 10
+numrho = 64
+# T is in Kelvin
+Tmin = 252
+Tmax = 1e4
+numT = 32
+# sie is in erg/g
+siemin = 1e12
+siemax = 1e16
+numsie = 32
+
+
+# titanium.dat
+matid = 2961
+name = titanium
+# These set the number of grid poitns per decade
+# for each variable. The default is 50 points.
+numrho/decade = 30
+numT/decade = 25
+numSie/decade = 15
+
+
+# steel.dat
+matid=4272
+rhomin = 1e-2
+Tmin = 1
+# These shrink lograithm of bounds
+# by a fraction of the total interval <= 1
+shrinklRhoBounds = 0.15
+shrinklTBounds = 0.15
+shrinkleBounds = 0.5
+)";
+
 // Parse a simple parameter file with
 // "#" denoting comments.
 class Params {
